% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/square_vect.R
\name{square_vect}
\alias{square_vect}
\title{squaring the input}
\usage{
square_vect(x)
}
\arguments{
\item{x}{}
}
\value{
vector, integer
}
\description{
squaring the input
}
\examples{
square_vect(c(1, 5, 10))
}

